refactor(M6 PR2): extract in-file nav (prompt) into controller/infile.rs
Move the bottom-prompt feature — go-to-line (`:`) + incremental search
(`/`, n/N): open, key handling, match navigation, refresh, and the source-line
jump — into a submodule. Cross-dispatch methods pub(super). Behavior-preserving.
mod.rs ~2294.
